+++ piccolo_api/session_auth/endpoints.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import os
 import typing as t
 import warnings
-from abc import ABCMeta, abstractproperty
+from abc import ABCMeta, abstractmethod
 from datetime import datetime, timedelta
 from json import JSONDecodeError
 
@@ -17,8 +17,9 @@ from starlette.responses import (
     PlainTextResponse,
     RedirectResponse,
 )
-from starlette.status import HTTP_303_SEE_OTHER
+from starlette.status import HTTP_303_SEE_OTHER, HTTP_401_UNAUTHORIZED
 
+from piccolo_api.mfa.provider import MFAProvider
 from piccolo_api.session_auth.tables import SessionsBase
 from piccolo_api.shared.auth.hooks import LoginHooks
 from piccolo_api.shared.auth.styles import Styles
@@ -39,23 +40,28 @@ LOGOUT_TEMPLATE_PATH = os.path.join(TEMPLATE_DIR, "session_logout.html")
 
 
 class SessionLogoutEndpoint(HTTPEndpoint, metaclass=ABCMeta):
-    @abstractproperty
+    @property
+    @abstractmethod
     def _session_table(self) -> t.Type[SessionsBase]:
         raise NotImplementedError
 
-    @abstractproperty
+    @property
+    @abstractmethod
     def _cookie_name(self) -> str:
         raise NotImplementedError
 
-    @abstractproperty
+    @property
+    @abstractmethod
     def _redirect_to(self) -> t.Optional[str]:
         raise NotImplementedError
 
-    @abstractproperty
+    @property
+    @abstractmethod
     def _logout_template(self) -> Template:
         raise NotImplementedError
 
-    @abstractproperty
+    @property
+    @abstractmethod
     def _styles(self) -> t.Optional[Styles]:
         raise NotImplementedError
 
@@ -86,7 +92,8 @@ class SessionLogoutEndpoint(HTTPEndpoint, metaclass=ABCMeta):
         cookie = request.cookies.get(self._cookie_name, None)
         if not cookie:
             raise HTTPException(
-                status_code=401, detail="The session cookie wasn't found."
+                status_code=HTTP_401_UNAUTHORIZED,
+                detail="The session cookie wasn't found.",
             )
         await self._session_table.remove_session(token=cookie)
 
@@ -102,54 +109,70 @@ class SessionLogoutEndpoint(HTTPEndpoint, metaclass=ABCMeta):
 
 
 class SessionLoginEndpoint(HTTPEndpoint, metaclass=ABCMeta):
-    @abstractproperty
+    @property
+    @abstractmethod
     def _auth_table(self) -> t.Type[BaseUser]:
         raise NotImplementedError
 
-    @abstractproperty
+    @property
+    @abstractmethod
     def _session_table(self) -> t.Type[SessionsBase]:
         raise NotImplementedError
 
-    @abstractproperty
+    @property
+    @abstractmethod
     def _session_expiry(self) -> timedelta:
         raise NotImplementedError
 
-    @abstractproperty
+    @property
+    @abstractmethod
     def _max_session_expiry(self) -> timedelta:
         raise NotImplementedError
 
-    @abstractproperty
+    @property
+    @abstractmethod
     def _cookie_name(self) -> str:
         raise NotImplementedError
 
-    @abstractproperty
+    @property
+    @abstractmethod
     def _redirect_to(self) -> t.Optional[str]:
         """
         Where to redirect to after login is successful.
         """
         raise NotImplementedError
 
-    @abstractproperty
+    @property
+    @abstractmethod
     def _production(self) -> bool:
         """
         If True, apply more stringent security.
         """
         raise NotImplementedError
 
-    @abstractproperty
+    @property
+    @abstractmethod
     def _login_template(self) -> Template:
         raise NotImplementedError
 
-    @abstractproperty
+    @property
+    @abstractmethod
     def _hooks(self) -> t.Optional[LoginHooks]:
         raise NotImplementedError
 
-    @abstractproperty
+    @property
+    @abstractmethod
     def _captcha(self) -> t.Optional[Captcha]:
         raise NotImplementedError
 
-    @abstractproperty
+    @property
+    @abstractmethod
     def _styles(self) -> t.Optional[Styles]:
+        raise NotImplementedError
+
+    @property
+    @abstractmethod
+    def _mfa_providers(self) -> t.Optional[t.Sequence[MFAProvider]]:
         raise NotImplementedError
 
     def _render_template(
@@ -182,11 +205,14 @@ class SessionLoginEndpoint(HTTPEndpoint, metaclass=ABCMeta):
     ) -> Response:
         if response_format == "html":
             return self._render_template(
-                request, template_context={"error": error}, status_code=401
+                request,
+                template_context={"error": error},
+                status_code=HTTP_401_UNAUTHORIZED,
             )
         else:
             return PlainTextResponse(
-                status_code=401, content=f"Login failed: {error}"
+                status_code=HTTP_401_UNAUTHORIZED,
+                content=f"Login failed: {error}",
             )
 
     async def get(self, request: Request) -> HTMLResponse:
@@ -203,8 +229,8 @@ class SessionLoginEndpoint(HTTPEndpoint, metaclass=ABCMeta):
             except JSONDecodeError:
                 body = await request.form()
 
-        username = body.get("username", None)
-        password = body.get("password", None)
+        username = body.get("username")
+        password = body.get("password")
         return_html = body.get("format") == "html"
 
         if (not username) or (not password):
@@ -239,7 +265,8 @@ class SessionLoginEndpoint(HTTPEndpoint, metaclass=ABCMeta):
                     )
                 else:
                     raise HTTPException(
-                        status_code=401, detail=validate_response
+                        status_code=HTTP_401_UNAUTHORIZED,
+                        detail=validate_response,
                     )
 
         # Attempt login
@@ -248,6 +275,109 @@ class SessionLoginEndpoint(HTTPEndpoint, metaclass=ABCMeta):
         )
 
         if user_id:
+            # Apply MFA
+            if mfa_providers := self._mfa_providers:
+                user = (
+                    await self._auth_table.objects()
+                    .where(self._auth_table.id == user_id)
+                    .first()
+                )
+
+                assert user is not None
+
+                if enrolled_mfa_providers := [
+                    mfa_provider
+                    for mfa_provider in mfa_providers
+                    if await mfa_provider.is_user_enrolled(user=user)
+                ]:
+                    mfa_code = body.get("mfa_code")
+
+                    if mfa_code is None:
+                        has_sent_code: t.List[bool] = []
+                        for mfa_provider in enrolled_mfa_providers:
+                            # Send the code (only used with things like email
+                            # and SMS MFA).
+                            has_sent_code.append(
+                                await mfa_provider.send_code(user=user)
+                            )
+
+                        message = "MFA code required"
+                        if any(has_sent_code):
+                            message += " (we sent you a code)"
+
+                        if return_html:
+                            return self._render_template(
+                                request,
+                                template_context={
+                                    "error": message,
+                                    "show_mfa_input": True,
+                                    "mfa_provider_names": [
+                                        mfa_provider.name
+                                        for mfa_provider in enrolled_mfa_providers  # noqa: E501
+                                    ],
+                                },
+                            )
+                        else:
+                            raise HTTPException(
+                                status_code=HTTP_401_UNAUTHORIZED,
+                                detail=message,
+                            )
+
+                    # Work out which MFA provider to use:
+                    if len(enrolled_mfa_providers) == 1:
+                        active_mfa_provider = enrolled_mfa_providers[0]
+                    else:
+                        mfa_provider_name = body.get("mfa_provider_name")
+
+                        if mfa_provider_name is None:
+                            raise HTTPException(
+                                status_code=HTTP_401_UNAUTHORIZED,
+                                detail="MFA provider must be specified",
+                            )
+
+                        filtered_mfa_providers = [
+                            i
+                            for i in enrolled_mfa_providers
+                            if i.name == mfa_provider_name
+                        ]
+
+                        if len(filtered_mfa_providers) == 0:
+                            raise HTTPException(
+                                status_code=HTTP_401_UNAUTHORIZED,
+                                detail="MFA provider not recognised.",
+                            )
+
+                        if len(filtered_mfa_providers) > 1:
+                            raise HTTPException(
+                                status_code=HTTP_401_UNAUTHORIZED,
+                                detail=(
+                                    "Multiple matching MFA providers found."
+                                ),
+                            )
+
+                        active_mfa_provider = filtered_mfa_providers[0]
+
+                    if not await active_mfa_provider.authenticate_user(
+                        user=user, code=mfa_code
+                    ):
+                        if return_html:
+                            return self._render_template(
+                                request,
+                                template_context={
+                                    "error": "MFA failed",
+                                    "show_mfa_input": True,
+                                    "mfa_provider_names": {
+                                        mfa_provider.name
+                                        for mfa_provider in enrolled_mfa_providers  # noqa: E501
+                                    },
+                                },
+                            )
+                        else:
+                            raise HTTPException(
+                                status_code=HTTP_401_UNAUTHORIZED,
+                                detail="MFA failed",
+                            )
+
             # Run login_success hooks
             if self._hooks and self._hooks.login_success:
                 hooks_response = await self._hooks.run_login_success(
@@ -280,7 +410,9 @@ class SessionLoginEndpoint(HTTPEndpoint, metaclass=ABCMeta):
                     },
                 )
             else:
-                raise HTTPException(status_code=401, detail="Login failed")
+                raise HTTPException(
+                    status_code=HTTP_401_UNAUTHORIZED, detail="Login failed"
+                )
 
         now = datetime.now()
         expiry_date = now + self._session_expiry
@@ -333,6 +465,7 @@ def session_login(
     hooks: t.Optional[LoginHooks] = None,
     captcha: t.Optional[Captcha] = None,
     styles: t.Optional[Styles] = None,
+    mfa_providers: t.Optional[t.Sequence[MFAProvider]] = None,
 ) -> t.Type[SessionLoginEndpoint]:
     """
     An endpoint for creating a user session.
@@ -372,6 +505,9 @@ def session_login(
         See :class:`Captcha <piccolo_api.shared.auth.captcha.Captcha>`.
     :param styles:
         Modify the appearance of the HTML template using CSS.
+    :param mfa_providers:
+        Add additional security to the login process using Multi-Factor
+        Authentication.
 
     """  # noqa: E501
     template_path = (
@@ -396,6 +532,7 @@ def session_login(
         _hooks = hooks
         _captcha = captcha
         _styles = styles or Styles()
+        _mfa_providers = mfa_providers
 
     return _SessionLoginEndpoint